setup des fichiers pour test de coverage. Migration des fonctions utils dans le module utils. Implémentation des tests unitaires pour les fonctions utils. Affichage des messages flash dans le template index. Création des fonctions utils pour le login.
