+++ server.py
@@ -1,18 +1,6 @@
 import json
 from flask import Flask,render_template,request,redirect,flash,url_for
-
-
-def loadClubs():
-    with open('clubs.json') as c:
-         listOfClubs = json.load(c)['clubs']
-         return listOfClubs
-
-
-def loadCompetitions():
-    with open('competitions.json') as comps:
-         listOfCompetitions = json.load(comps)['competitions']
-         return listOfCompetitions
-
+from utils import loadClubs, loadCompetitions, getClubByEmail
 
 app = Flask(__name__)
 app.secret_key = 'something_special'
@@ -26,8 +14,16 @@ def index():
 
 @app.route('/showSummary',methods=['POST'])
 def showSummary():
-    club = [club for club in clubs if club['email'] == request.form['email']][0]
-    return render_template('welcome.html',club=club,competitions=competitions)
+    if clubs is None or competitions is None:
+        flash("Error loading clubs or competitions data.")
+        return redirect(url_for('index'))
+    
+    club = getClubByEmail(request.form['email'], clubs)
+    if club:
+        return render_template('welcome.html',club=club,competitions=competitions)
+    else:
+        flash("Unfortunately, the email you entered was not found.")
+        return redirect(url_for('index'))
 
 
 @app.route('/book/<competition>/<club>')
